requirements.txt
Streamlit
google-generativeai
python-dotenv